likes

+++ backend/app/controllers/api/posts_controller.rb
@@ -1,11 +1,11 @@
 # rubocop:disable Style/ClassAndModuleChildren
 class Api::PostsController < ApplicationController
   def index
-    @posts = Post.order(created_at: :desc).limit(100)
+    @posts = Post.includes(:author, :likers).order(created_at: :desc).limit(100)
   end
 
   def create
-    require_signed_in
+    return if require_signed_in
     @post = Post.new(post_params.merge(author_id: current_user.id))
     if @post.save
       render :show
@@ -15,12 +15,12 @@ class Api::PostsController < ApplicationController
   end
 
   def show
-    @post = Post.find_by(id: params[:id])
+    @post = Post.includes(:author, :likers).find_by(id: params[:id])
     render :show
   end
 
   def update
-    require_signed_in
+    return if require_signed_in
     @post = Post.find_by(id: params[:id])
     render json: { errors: ["Cannot edit other user's post"] } if @post.author_id != current_user.id
     if @post.update(post_params)
@@ -31,7 +31,7 @@ class Api::PostsController < ApplicationController
   end
 
   def destroy
-    require_signed_in
+    return if require_signed_in
     @post = Post.find_by(id: params[:id])
     render json: { errors: ["Cannot delete other user's post"] } if @post.author_id != current_user.id
     if @post.destroy
